[dev] feat: add physics-aware full S-matrix surrogate workflow
- implement reciprocal complex S-matrix models with Fourier/RBF features
- add guarded fine-tuning, residual correction heads, and physics diagnostics
- add the NB06 training, evaluation, persistence, and benchmark workflow
- support derived insertion-loss benchmarks and diagnostic metrics
- store curve targets as float32 and expand model coverage with unit tests

+++ sparam-surrogate/src/sparam_surrogate/data/curve_dataset.py
@@ -205,7 +205,7 @@ class CurveDataset:
         for row in records:
             frequencies_ghz, targets = curve_loader.load_curve(row)
             frequencies_ghz = np.asarray(frequencies_ghz, dtype=float)
-            targets = np.asarray(targets, dtype=float)
+            targets = np.asarray(targets, dtype=np.float32)
             expected_shape = (len(frequencies_ghz), target_width)
             if (
                 frequencies_ghz.ndim != 1
